tests: #148 step 8 — move _agent_targets_stage helper from test_feedback_keepalive to test_lib
